chore: Remove GNOME Flatpaks
We'll have to figure this out another day to replace them with equivalents that **verified Qt applications**.  KDE has a Flatpak repo that contains a lot of equivalents, but that would be messy to include another Flatpak repo.

What has been removed:

-GNOME Calendar
- GNOME Characters
- GNOME Connections
- GNOME Logs
- GNOME Maps

I left in Mission Center.  It is up to you if you want to keep that or not in the future.